requirements.txt
pydantic>=2.6,<3
python-binance>=1.0.19
pandas>=2.0
ta>=0.11.0
python-dotenv>=1.0
schedule>=1.2.0
requests>=2.31
rich>=13.7
openai>=1.0

